Working with logic

+++ app/controllers/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controllers;
 
+use App\Connect\DataBase;
 use App\Models\Profile;
 use App\Models\User;
 use App\Views\RenderView;
@@ -19,7 +20,6 @@ class UserController
                 $profile = Profile::getProfile($query, $user->getId());
                 if (!empty($user)) {
                     if (!empty($profile)) {
-                        $profile->getUserId();
                         $_SESSION['isProfileSave'] = $profile->getStatusIsProfileSave();
                         $_SESSION['userId'] = $profile->getUserId();
                     }
@@ -30,12 +30,10 @@ class UserController
                 }
                 RenderView::getRenderViewObject()->renderTemplate($arr, $user);
             } catch (\Exception $e) {
-                $arr['error'] = $e->getMessage();
+                $arr['error'] = 'Ошибка! ' . $e->getMessage();
                 RenderView::getRenderViewObject()->renderTemplate($arr);
-                die;
             }
-        }
-        if (isset($_SESSION['auth'])) {
+        } elseif (isset($_SESSION['auth'])) {
             header('Location: /content/profile/list');
             die;
         } else {
@@ -53,48 +51,56 @@ class UserController
     {
         if (!empty($_POST)) {
             try {
-                $success = User::writeDataBase('INSERT INTO `user` (email, password, nickname) VALUES (?, ?, ?)', $_POST);
+                $query = 'INSERT INTO `user` (email, password, nickname) VALUES (?, ?, ?)';
+                $success = User::writeDataBase($query, $_POST);
                 if ($success) {
-                    header('Location: /auth/user/log');
+                    $_SESSION['auth'] = true;
+                    $_SESSION['flash'] = 'Вы успешно зарегестрировались';
+                    header('Location: /content/profile/list');
                     die;
                 }
             } catch (\Exception $e) {
-                $arr['error'] = $e->getMessage();
+                $arr['error'] = 'Ошибка! ' . $e->getMessage();
                 RenderView::getRenderViewObject()->renderTemplate($arr);
-                die;
             }
+        } else {
+            RenderView::getRenderViewObject()->renderTemplate($arr);
         }
-        RenderView::getRenderViewObject()->renderTemplate($arr);
     }
 
     public function deleteAccount($arr)
     {
-        if (!empty($_POST)) {
-            $query = 'SELECT * FROM `user` WHERE id = ?';
-            try {
-                $user = User::getUser($query);
-                var_dump($_SESSION);
-                var_dump($user);
-                if (password_verify($_POST['password'], $user->getPassword())) {
-                    $query = 'DELETE FROM `user` WHERE id = ?';
-                    User::deleteUser($query);
-                    $query = 'DELETE FROM `profile` WHERE user_id = ?';
-                    Profile::deleteProfile($query);
-                    session_destroy();
-                    header('Location: /content/profile/list');
-                    die;
-                } else {
-                    $_SESSION['flash'] = 'Пароль неверный';
-                    header('Location: /content/profile/list');
+        if (isset($_SESSION['auth'])) {
+            if (!empty($_POST)) {
+                $query = 'SELECT * FROM `user` WHERE id = ?';
+                try {
+                    $user = User::getUser($query);
+                    var_dump($_SESSION);
+                    var_dump($user);
+                    if (password_verify($_POST['password'], $user->getPassword())) {
+                        $query = 'DELETE FROM `user` WHERE id = ?';
+                        User::deleteUser($query);
+                        $query = 'DELETE FROM `profile` WHERE user_id = ?';
+                        Profile::deleteProfile($query);
+                        session_destroy();
+                        header('Location: /content/profile/list');
+                        die;
+                    } else {
+                        $_SESSION['flash'] = 'Пароль неверный';
+                        header('Location: /content/profile/list');
+                        die;
+                    }
+                } catch (\Exception $e) {
+                    $arr['error'] = $e->getMessage();
+                    RenderView::getRenderViewObject()->renderTemplate($arr);
                     die;
                 }
-            } catch (\Exception $e) {
-                $arr['error'] = $e->getMessage();
+            } else {
                 RenderView::getRenderViewObject()->renderTemplate($arr);
-                die;
             }
         } else {
-            RenderView::getRenderViewObject()->renderTemplate($arr);
+            header('Location: /auth/user/log');
+            die;
         }
     }
 }
